middleman/requirements.txt
# Python dependencies
uwsgi
Flask
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
requests
Flask-Caching
jsonschema
psycopg2
Flask-BasicAuth